Add addrs for EGS non-Denuvo ver (EGS-WW-unprot)
Also added AltCheckData for UWP 1.0.4.0 release

+++ DQXI-SDK/pch.h
@@ -35,7 +35,7 @@ struct GameAddresses
   // If neither CheckData/CheckDataJP match, check one other offset
   // (Denuvo can change timestamp addr/value, even though Denuvo/non-Denuvo EXE pretty much match)
   uintptr_t AltCheckDataAddr;
-  unsigned int AltCheckData;
+  unsigned long long AltCheckData;
 
   // Variables
   uintptr_t GUObjectArray; // aka GObjects
